Functionality Update
Everything should be functional now
