--- Assets/Scenes/Joseph Seo/Save System.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 534602a8ef570c54497b02a55a5acf3b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
