--- Assets/Scenes/Joseph Seo/Save System/DeserializationTypeRemapBinder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 50be4f5b8078be345a78ed858426ada2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
