--- Assets/Scenes/Joseph Seo/Save System/GameSaver.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class GameSaver : MonoBehaviour
-{
-    private static GameSaver instance;
-    private static SaveSystem saveSystem;
-
-    private void Awake() {
-        if (instance == null) {
-            instance = this;
-            saveSystem = new SaveSystem();
-        } else {
-            // Destroy();
-        }
-
-    }
-}
