--- Assets/Scenes/Joseph Seo/Save System/ISavable.cs
@@ -1,5 +1,0 @@
-public interface ISavable 
-{
-    void Save();
-    void Load(SaveProfile profile);
-}
