--- Assets/Scenes/Joseph Seo/Save System/NavaScripts.asmdef
@@ -1,3 +1,0 @@
-{
-	"name": "NavaScripts"
-}
