--- Assets/Scenes/Joseph Seo/Save System/SaveMainMenu.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9fcb8d5dc8175b34cb08d09d15cdeb48
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
