--- Assets/Scenes/Joseph Seo/Save System/SaveSceneLoad.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 318c887fedd432e4da074768d8fe2a12
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
