--- Assets/Scenes/Joseph Seo/Save System/SaveSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca93229e1a1a4dd468ad85af3fa01817
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
