--- Assets/Scenes/Joseph Seo/Save System/SceneInitializer.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class SceneInitializer : MonoBehaviour
-{
-    private void Start() {
-        SaveSystem.Current.Load();
-    }
-
-    void OnApplicationQuit() {
-        SaveSystem.SaveGame();
-    }
-}
