--- Assets/Scenes/Joseph Seo/Save System/SceneInitializer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 779ad73604fd8e14c90663701691e719
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 5
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
