+++ Assets/Scripts/SaveLoadManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9924264e5c3ccde458fed0f3634386e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
