+++ Assets/Scripts/SaveSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bdfd6c02c78774f4fb92c23bc41c225c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
