+++ Assets/Scripts/SaveSystem/GameSaver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e3b5463bd581893478c23c4953edbce8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
