+++ Assets/Scripts/SaveSystem/ISavable.cs
@@ -1,0 +1,5 @@
+public interface ISavable 
+{
+    void Save();
+    void Load(SaveProfile profile);
+}
