+++ Assets/Scripts/SaveSystem/ISavable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69bc60445986c7744ab8d38a35fc22c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
