+++ Assets/Scripts/SaveSystem/NavaScripts.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "NavaScripts"
+}
