+++ Assets/Scripts/SaveSystem/NavaScripts.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 47a2db908e23c77409751e6a48192c0b
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
