+++ Assets/Scripts/SaveSystem/SaveProfile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 174fbcc94ab4de54085e33c56013c909
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
