+++ Assets/Scripts/SaveSystem/SerializableSaveProfile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ff2e2d0f1f8d624093432638a033914
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
